refactor(model): Move TileAction interface to models.interfaces package
