+++ src/main/java/edu/ntnu/idi/idatt/model/SomeOtherAction.java
@@ -1,5 +1,7 @@
 package edu.ntnu.idi.idatt.model;
 
+import edu.ntnu.idi.idatt.model.interfaces.TileAction;
+
 public class SomeOtherAction implements TileAction {
 
   public void perform(Player player, Board board) {
