--- src/main/java/edu/ntnu/idi/idatt/model/TileAction.java
@@ -1,6 +1,0 @@
-package edu.ntnu.idi.idatt.model;
-
-public interface TileAction {
-
-  public void perform(Player player, Board board);
-}
